feat: pss-cln-work-fold-service-001 - Fold composed Work root into owner-private internal
