+++ cmd/packagetargetmanifestcheck/owners.go
@@ -284,6 +284,7 @@ var nestedOwnerMoveRules = map[string][]nestedPathRule{
 		{exact: "token_transformer", prefix: "token_transformer/", dest: "factory_runtime/internal/services/orchestration"},
 	},
 	"work": {
+		{exact: "internal/service", prefix: "internal/service/", dest: "work/internal"},
 		{exact: "materialize", prefix: "materialize/", dest: "work/internal/services/content_materialization"},
 		{exact: "stateaccessrecordings", prefix: "stateaccessrecordings/", dest: "work/internal/services/state_access"},
 		{exact: "testdata", prefix: "testdata/", dest: "work/internal"},
